refactor(i18n): abstract into repo package
