+++ apps/web/package.json
@@ -24,6 +24,7 @@
     "@repo/design-system": "workspace:*",
     "@repo/email": "workspace:*",
     "@repo/env": "workspace:*",
+    "@repo/i18n": "workspace:*",
     "@repo/next-config": "workspace:*",
     "@repo/observability": "workspace:*",
     "@repo/og-utils": "workspace:*",
@@ -38,7 +39,6 @@
     "motion": "^11.13.4",
     "negotiator": "^1.0.0",
     "next": "15.1.0",
-    "next-intl": "^3.25.3",
     "react": "19.0.0",
     "react-dom": "19.0.0",
     "react-hook-form": "^7.53.2",
