+++ apps/web/src/app/[locale]/(pages)/[slug]/route-data.tsx
@@ -1,8 +1,8 @@
 import { stripTrailingSlash } from '@repo/design-system/lib/utils/string'
+import { type Locale } from '@repo/i18n/types'
 import { notFound } from 'next/navigation'
 import * as React from 'react'
 import { getPage } from '../api'
-import { type Locale } from '~/app/i18n/types'
 
 // slug is empty string when on the homepage
 export const getRouteData = React.cache(async function getRouteData(
