+++ apps/web/src/app/[locale]/(pages)/api.ts
@@ -1,3 +1,5 @@
+import { i18n } from '@repo/i18n/config'
+import { type Locale } from '@repo/i18n/types'
 import { tina } from '~/app/cms/client'
 import {
 	type ContentQueryQuery,
@@ -5,8 +7,6 @@ import {
 	type Wiki,
 	type Work,
 } from '~/app/cms/types'
-import { i18n } from '~/app/i18n/config'
-import { type Locale } from '~/app/i18n/types'
 
 export async function getPages() {
 	const pageDate = await tina.queries.pageConnection()
