+++ apps/web/src/app/[locale]/blog/[...slug]/route-data.tsx
@@ -1,7 +1,7 @@
+import { type Locale } from '@repo/i18n/types'
 import { notFound } from 'next/navigation'
 import * as React from 'react'
 import { getBlogPost } from '../api'
-import { type Locale } from '~/app/i18n/types'
 
 export const getRouteData = React.cache(async function getRouteData(
 	slug: string,
