+++ apps/web/src/app/[locale]/blog/utils.ts
@@ -1,10 +1,10 @@
 import { type Article } from '@repo/design-system/components/article-overview'
 import { nonNullable } from '@repo/design-system/lib/utils/array'
 import { type PartialDeep } from '@repo/design-system/lib/utils/types'
+import { i18n } from '@repo/i18n/config'
 import uniqueId from 'lodash/uniqueId'
 import { matchSorter, rankings as matchSorterRankings } from 'match-sorter'
 import { type Blog } from '~/app/cms/types'
-import { i18n } from '~/app/i18n/config'
 
 const BlogPath = 'blog'
