+++ apps/web/src/app/[locale]/contact/blocks/chat.tsx
@@ -4,7 +4,7 @@ import { cx } from '@nerdfish/utils'
 import { Button, useMediaQuery } from '@repo/design-system/components/ui'
 import { useLockBody } from '@repo/design-system/hooks/use-body-lock'
 import { BotIcon, XIcon } from '@repo/design-system/lib/icons'
-import { useTranslations } from 'next-intl'
+import { useTranslations } from '@repo/i18n/client'
 import * as React from 'react'
 import { Chat } from '../components/chat'
 import { type Block, type PageBlocksChatbot } from '~/app/cms/types'
