+++ apps/web/src/app/[locale]/layout.tsx
@@ -1,13 +1,13 @@
 import { fonts } from '@repo/design-system/lib/fonts'
-import { NextIntlClientProvider } from 'next-intl'
+import { NextIntlClientProvider } from '@repo/i18n/client'
+import { i18n } from '@repo/i18n/config'
+import { type WithLocale } from '@repo/i18n/types'
 import { getMessages, setRequestLocale } from 'next-intl/server'
 import * as React from 'react'
 import { AppProviders } from '../app-providers'
 import { getGlobalData } from '../cms/api'
 import { GlobalProvider } from '../global-provider'
 import { SiteLayout } from '../layout/site-layout'
-import { i18n } from '~/app/i18n/config'
-import { type WithLocale } from '~/app/i18n/types'
 
 import '~/app/theme/styles/app.css'
 import '~/app/theme/styles/prose.css'
