+++ apps/web/src/app/[locale]/wiki/blocks/wiki-overview.tsx
@@ -30,8 +30,8 @@ import {
 } from '@repo/design-system/components/ui'
 import { BookIcon, PlusIcon, SearchIcon } from '@repo/design-system/lib/icons'
 import { nonNullable } from '@repo/design-system/lib/utils/array'
+import { useTranslations } from '@repo/i18n/client'
 import Image from 'next/image'
-import { useTranslations } from 'next-intl'
 import * as React from 'react'
 import { tinaField } from 'tinacms/dist/react'
 import { filterWiki, getWikiPath } from '../utils'
