+++ apps/web/src/app/[locale]/work/[...slug]/page.tsx
@@ -1,3 +1,4 @@
+import { type WithLocale } from '@repo/i18n/types'
 import { pageParams } from '@repo/og-utils/zod-params'
 import { createMetadata } from '@repo/seo/metadata'
 import { type Metadata } from 'next'
@@ -8,7 +9,6 @@ import { WorkContent } from '../components/work-content'
 import { WorkPreview } from '../components/work-preview'
 import { getWorkPath } from '../utils'
 import { getRouteData } from './route-data'
-import { type WithLocale } from '~/app/i18n/types'
 
 export async function generateMetadata(props: {
 	params: Promise<WithLocale<{ slug: string[] }>>
