+++ apps/web/src/app/[locale]/work/api.ts
@@ -1,5 +1,5 @@
+import { type Locale } from '@repo/i18n/types'
 import { tina } from '~/app/cms/client'
-import { type Locale } from '~/app/i18n/types'
 
 export async function getWorks({ locale }: { locale?: Locale } = {}) {
 	const workListData = await tina.queries.workConnection()
