+++ apps/web/src/app/cms/api.ts
@@ -1,5 +1,5 @@
-import { i18n } from '../i18n/config'
-import { type Locale } from '../i18n/types'
+import { i18n } from '@repo/i18n/config'
+import { type Locale } from '@repo/i18n/types'
 import tina from '~/tina/__generated__/client'
 
 export async function getGlobalData(locale: Locale = i18n.defaultLocale) {
