+++ apps/web/src/app/cms/blocks/features.tsx
@@ -10,11 +10,11 @@ import {
 } from '@repo/design-system/components/section'
 import { Grid, H3 } from '@repo/design-system/components/ui'
 import { ArrowRight } from '@repo/design-system/lib/icons'
+import { useTranslations } from '@repo/i18n/client'
 import { camelCase, startCase } from 'lodash'
 import * as Icons from 'lucide-react'
 import { useInView } from 'motion/react'
 import Link from 'next/link'
-import { useTranslations } from 'next-intl'
 import * as React from 'react'
 import { tinaField } from 'tinacms/dist/react'
 import { getPagePath } from '~/app/[locale]/(pages)/utils'
