+++ apps/web/src/middleware.ts
@@ -1,5 +1,4 @@
-import createMiddleware from 'next-intl/middleware'
-import { i18n } from './app/i18n/config'
+import { i18nMiddleware } from '@repo/i18n/middleware'
 
 export const config = {
 	/*
@@ -17,13 +16,4 @@ export const config = {
 	],
 }
 
-export default createMiddleware({
-	locales: i18n.locales,
-	defaultLocale: i18n.defaultLocale,
-	localePrefix: 'as-needed',
-	localeDetection: true,
-	localeCookie: {
-		secure: true,
-		sameSite: 'strict',
-	},
-})
+export default i18nMiddleware
