+++ packages/i18n/client.ts
@@ -1,0 +1,1 @@
+export * from 'next-intl'
