+++ packages/i18n/config.ts
