+++ packages/i18n/global.d.ts
