+++ packages/i18n/package.json
@@ -1,20 +1,22 @@
 {
-  "name": "@repo/next-config",
+  "name": "@repo/i18n",
   "version": "0.0.0",
   "private": true,
   "scripts": {
     "clean": "git clean -xdf .cache .turbo dist node_modules"
   },
   "dependencies": {
-    "@next/bundle-analyzer": "15.1.0",
-    "@vercel/toolbar": "^0.1.27",
-    "next-secure-headers": "^2.2.0"
+    "@repo/design-system": "workspace:*",
+    "next": "15.1.0",
+    "next-intl": "^3.25.3",
+    "react": "19.0.0",
+    "react-dom": "19.0.0",
+    "zod": "^3.23.8"
   },
   "devDependencies": {
     "@nerdfish/config": "^1.6.1",
     "@repo/env": "workspace:*",
     "next": "15.1.0",
-    "react": "19.0.0",
-    "react-dom": "19.0.0"
+    "typescript": "^5.7.2"
   }
 }
