+++ packages/i18n/plugin.ts
@@ -1,0 +1,9 @@
+import { type NextConfig } from 'next'
+import withNextIntl from 'next-intl/plugin'
+
+export const withTranslations = (
+	requestConfigPath: string,
+	conf: NextConfig,
+) => {
+	return withNextIntl(requestConfigPath)(conf)
+}
