+++ packages/i18n/server.ts
@@ -1,0 +1,5 @@
+export {
+	getMessages,
+	setRequestLocale,
+	getRequestConfig,
+} from 'next-intl/server'
