+++ packages/i18n/tsconfig.json
@@ -1,0 +1,8 @@
+{
+	"extends": "@nerdfish/config/typescript",
+	"compilerOptions": {
+		"baseUrl": "."
+	},
+	"include": ["**/*.ts", "**/*.tsx"],
+	"exclude": ["node_modules"]
+}
